Move SpringBootBaseIntegration class to different package Reset DB before every test
