+++ src/src/test/java/common/SpringBootBaseIntegration.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package runner;
+package common;
 
 import com.liquorexchange.LiquorExchangeApplication;
 import com.liquorexchange.db.DatabaseLoader;
@@ -12,6 +12,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.test.context.junit4.SpringRunner;
 
 /**
@@ -19,21 +21,21 @@ import org.springframework.test.context.junit4.SpringRunner;
  * @author divad_000
  */
 @RunWith(SpringRunner.class)
+@Configuration
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 public class SpringBootBaseIntegration {
 
-    @LocalServerPort
-    protected int port;
-
-    @Autowired
-    DatabaseLoader dl;
+    protected DatabaseLoader databaseLoader;
 
     private static boolean appStarted = false;
+    private static ConfigurableApplicationContext context;
 
     public SpringBootBaseIntegration() {
         if (appStarted == false) {
-            SpringApplication.run(LiquorExchangeApplication.class, new String[]{});
+            context = SpringApplication.run(LiquorExchangeApplication.class, new String[]{});
             appStarted = true;
         }
+
+        this.databaseLoader = context.getBean(DatabaseLoader.class);
     }
 }
