+++ src/src/test/java/seleniumgluecode/login.java
@@ -13,7 +13,7 @@ import org.junit.Ignore;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import runner.SpringBootBaseIntegration;
+import common.SpringBootBaseIntegration;
 
 /**
  *
